+++ packages/broker/src/core/logger/BrokerLogger.types.ts
@@ -17,7 +17,6 @@ export type BrokerLogEvent =
   // Message routing
   | 'handler.failed'
   | 'unicast.multiple_handlers'
-  | 'request.history_deprecated'
   // Hooks
   | 'hook.after_send.failed'
   | 'hook.failed'